Simple_AI_Agent/pyproject.toml
[project]
name = "simple-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "transformers>=4.56.2",
]
